Move test accordin to new structure
