+++ src/tqec/circuit/operations/operation_test.py
@@ -10,8 +10,11 @@ from tqec.circuit.operations.operation import (
     make_detector,
     make_observable,
     make_shift_coords,
+    observable_qubits_from_template,
 )
 from tqec.exceptions import TQECException
+from tqec.plaquette.library import XXXXMemoryPlaquette, ZZZZMemoryPlaquette
+from tqec.templates import RawRectangleTemplate
 
 _qubits_examples = [
     cirq.GridQubit(0, 0),
@@ -122,3 +125,27 @@ def test_relative_measurement_record_duplicated_measurement_data():
                 RelativeMeasurementData(origin, -1),
             ],
         )
+
+
+def test_raw_rectangle_default_observable_qubits():
+    template = RawRectangleTemplate(
+        [
+            [0, 1, 0, 1],
+            [1, 0, 1, 0],
+            [0, 1, 0, 1],
+            [1, 0, 1, 0],
+        ]
+    )
+    plaquettes = [
+        XXXXMemoryPlaquette([1, 2, 3, 4, 5, 6, 7, 8]),
+        ZZZZMemoryPlaquette([1, 3, 4, 5, 6, 8]),
+    ]
+    obs = observable_qubits_from_template(template, plaquettes)
+    result = [
+        (cirq.GridQubit(3, -1), 0),
+        (cirq.GridQubit(3, 1), 0),
+        (cirq.GridQubit(3, 3), 0),
+        (cirq.GridQubit(3, 5), 0),
+        (cirq.GridQubit(3, 7), 0),
+    ]
+    assert sorted(obs, key=lambda t: t[0].col == result)
